_building/requirements.txt
semantic_version~=2.8.5
jinja2~=2.11.2
PyYAML~=5.3.1


sphinx~=3.3.1
